+++ src/Atc.Cosmos.EventStore/EventStoreClientOptions.cs
@@ -10,30 +10,6 @@ public class EventStoreClientOptions
     public const string EmulatorEndpoint = "https://localhost:8081/";
     public const string EmulatorAuthKey = "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
 
-    [Obsolete("Call UseCosmosEmulator instead.")]
-    public const string CosmosEmulatorConnectionString = "AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
-
-#pragma warning disable CS0618 // Type or member is obsolete
-    private string? connectionString = CosmosEmulatorConnectionString;
-#pragma warning restore CS0618 // Type or member is obsolete
-
-    [Obsolete("Call UseCredentials instead.")]
-    public string? ConnectionString
-    {
-        get
-        {
-            return connectionString;
-        }
-
-        set
-        {
-            connectionString = value;
-            AuthKey = null;
-            Endpoint = null;
-            Credential = null;
-        }
-    }
-
     public string EventStoreDatabaseId { get; set; } = "EventStore";
 
     public string EventStoreContainerId { get; set; } = "event-store";
@@ -60,6 +36,11 @@ public class EventStoreClientOptions
 
     public TokenCredential? Credential { get; private set; }
 
+    public bool AllowAnyServerCertificate { get; private set; }
+
+    public EventStoreClientOptions()
+        => UseCosmosEmulator();
+
     /// <summary>
     /// Configure event store to use <seealso cref="TokenCredential"/> instead <see cref="AuthKey"/>.
     /// </summary>
@@ -73,13 +54,10 @@ public class EventStoreClientOptions
         Credential = credentials ?? throw new ArgumentNullException(nameof(credentials));
         Endpoint = endpoint ?? throw new ArgumentNullException(nameof(endpoint));
         AuthKey = null;
-#pragma warning disable CS0618 // Type or member is obsolete
-        connectionString = null;
-#pragma warning restore CS0618 // Type or member is obsolete
     }
 
     /// <summary>
-    /// Configure event storte to use AuthKey when connecting to cosmos db.
+    /// Configure event store to use AuthKey when connecting to cosmos db.
     /// </summary>
     /// <param name="endpoint">Cosmos account endpoint.</param>
     /// <param name="authKey">Authorization key to connect with.</param>
@@ -91,21 +69,20 @@ public class EventStoreClientOptions
         Credential = null;
         Endpoint = endpoint ?? throw new ArgumentNullException(nameof(endpoint));
         AuthKey = authKey ?? throw new ArgumentNullException(nameof(authKey));
-#pragma warning disable CS0618 // Type or member is obsolete
-        connectionString = null;
-#pragma warning restore CS0618 // Type or member is obsolete
     }
 
     /// <summary>
     /// Configure event store to use cosmos emulator.
     /// </summary>
-    public void UseCosmosEmulator()
+    /// <param name="endpoint">Optional custom cosmos emulator endpoint.</param>
+    /// <param name="allowAnyServerCertificate">Optionally configure cosmos client to accept any server certificate.</param>
+    public void UseCosmosEmulator(
+        string endpoint = EmulatorEndpoint,
+        bool allowAnyServerCertificate = false)
     {
         Credential = null;
-        Endpoint = EmulatorEndpoint;
+        Endpoint = endpoint;
         AuthKey = EmulatorAuthKey;
-#pragma warning disable CS0618 // Type or member is obsolete
-        connectionString = null;
-#pragma warning restore CS0618 // Type or member is obsolete
+        AllowAnyServerCertificate = allowAnyServerCertificate;
     }
 }